feat: Track and display listening time by hour
Introduces a new `listeningTimeByHour` field to the `ListeningStats` type. This change tracks the amount of time a user spends listening to the radio, broken down by the hour of the day.

Additionally, it integrates a new `ListeningDNA` component (imported as `Contact` in some files) that leverages Gemini to analyze user listening habits, including identifying top genres based on liked songs and requests. This component will provide more personalized insights to users.